Export links to individual files
