+++ chainercv/links/model/deeplab/__init__.py
@@ -1,6 +1,6 @@
+from chainercv.links.model.deeplab.aspp import SeparableASPP  # NOQA
 from chainercv.links.model.deeplab.deeplab_v3_plus import Decoder  # NOQA
 from chainercv.links.model.deeplab.deeplab_v3_plus import DeepLabV3plus  # NOQA
 from chainercv.links.model.deeplab.deeplab_v3_plus import DeepLabV3plusXception65  # NOQA
-from chainercv.links.model.deeplab.deeplab_v3_plus import SeparableASPP  # NOQA
-from chainercv.links.model.deeplab.deeplab_v3_plus import Xception65  # NOQA
-from chainercv.links.model.deeplab.deeplab_v3_plus import XceptionBlock  # NOQA
+from chainercv.links.model.deeplab.xception import Xception65  # NOQA
+from chainercv.links.model.deeplab.xception import XceptionBlock  # NOQA
